pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lovelush-backend"
version = "0.1.0"
description = "Backend service for Lovelush application"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "motor==3.3.2",
    "pydantic[email]==2.5.0",
    "pydantic-settings==2.1.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "pymongo==4.6.0",
    "bcrypt==4.3.0",
    "colorlog==6.6.0",
    "python-telegram-bot==22.3",
    "aiohttp==3.12.15",
    "pusher==3.3.2",
    "boto3==1.40.21",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "pycln",
    "pre-commit",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
